- GTM_INLINE will make sure a function gets inlined, and provides a consistent   way for all GTM code to do it.
- Added GTMDebugThreadValidation to allow you to enforce the fact that your
  code must run in the main thread in DEBUG builds.

- Updated some internals of the iPhone unittesting so it doesn't double print
  the test descriptions, file names, or lines numbers of a test failure line.
  Also includes the test names in the error output.

- Changed the xcconfigs so that know it's easier to set different settings at
  the different levels and not accidentally overwrite settings set at lower
  levels in the "settings collapse". Also tightened up warnings significantly.

- Changed how gtm_unitTestExposedBindingsTestValues works. If you have an
  implementation of gtm_unitTestExposedBindingsTestValues in your own code
  you will need to update to the new way of calling. See implementations in
  GTMNSObject+BindingUnitTesting.m for details.

- Added support for grabbing the build number for a particular OS in
  GTMSystemVersion and easily comparing it to known build numbers, and switched 
  some types from in GTMSystemVersion from "int" to SInt32 to make 64 bit work 
  better.
   
- Added support for SnowLeopard (10A96). We build cleanly with the 10.6 SDKs and
  all radar checks were updated accordingly. Build All script was also updated
  to build on SnowLeopard if you have the SDK available.
  
- Turned off building ppc64 GTM because the SnowLeopard SDK currently
  doesn't have ppc64 support, so SenTestCase isn't defined. This makes it
  impossible to build the ppc64 10.5 config on SnowLeopard. We have left the
  setting in the xcconfig for those of you who need it, but have disabled
  it in the GTM project settings.
  
- Turned on stack smashing protection on the debug builds for all Leopard 
  and above.

- Added ability to easily do leak checking by defining the GTM_ENABLE_LEAKS
  environment variable. It isn't on by default because several of Apple's
  frameworks leak. You can work around these false positives by using the
  GTM_LEAKS_SYMBOLS_TO_IGNORE environment variable. Also if you turn on leaks
  make sure to turn off zombies by defining the GTM_DISABLE_ZOMBIES variable,
  otherwise every memory allocation you do will look like a leak.

+++ trunk/Foundation/GTMStackTrace.h
@@ -17,6 +17,7 @@
 //
 
 #include <CoreFoundation/CoreFoundation.h>
+#import "GTMDefines.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -66,7 +67,7 @@ NSString *GTMStackTrace(void);
 // Returns:
 //   The number of program counters actually added to outPcs.
 //
-int GTMGetStackProgramCounters(void *outPcs[], int count);
+NSUInteger GTMGetStackProgramCounters(void *outPcs[], NSUInteger count);
 
 // Returns an array of GTMAddressDescriptors from the current thread's stack.
 // *** You should probably use GTMStackTrace() instead of this function ***
@@ -84,8 +85,8 @@ int GTMGetStackProgramCounters(void *outPcs[], int count);
 // Returns:
 //   The number of program counters actually added to outPcs.
 //
-int GTMGetStackAddressDescriptors(struct GTMAddressDescriptor outDescs[], 
-                                  int count);
+NSUInteger GTMGetStackAddressDescriptors(struct GTMAddressDescriptor outDescs[], 
+                                         NSUInteger count);
 #ifdef __cplusplus
 }
 #endif
